Set NUM_OF_SEEDS to 2 instead of 3 and update KATs
The number of seeds (NUM_OF_SEEDS) was previously set to 3 here, only
for compatibility with the BIKE Reference code (at that time) that
generates 3 seeds:
https://bikesuite.org/files/v4.0/Reference_Implementation.2020.05.03.1.zip

The actual number of seeds that are required for BIKE, and used in the
protocol, is 2.
BIKE team fixed this issue on 06/03/2020, and uploaded the new version of
Reference code:
https://bikesuite.org/files/v4.0/Reference_Implementation.2020.06.03.1.zip

To be compatible with the revised Reference code, we now set the number
of seeds to 2 (instead of 3) in this code package, and update the KATs accordingly.

BIKE team fixed several more inconsistencies which were responsible
for the difference between the KATs of this code and the Reference code.
The changes they made are listed in:
https://bikesuite.org/files/v4.0/change-log.2020.06.03.1.txt

The reasons for different KATs were explained in tests/kats/Readme.md.
Since now the KATs are compatible with the Reference code KATs, the
Readme file is updated to reflect the new state.

Signed-off-by: Dusan Kostic <dusan.kostic@epfl.ch>