Runtime code path selection based on CPU capabilities (#8)
* Runtime code path selection based on CPU capabilities

Before this commit the code could be compiled for different
platforms that support a specified x86 instruction set (PCLMUL,
VPCLMUL, AVX2, AVX512). With this commit we change the code such
that all the versions are compiled in a single binary and the
selection of the correct functions is done in runtime based
on the CPU capabilities. This means that the compiled binary
can run on any CPU and automatically choose the most optimized
code that can be executed on the CPU.

Signed-off-by: Dusan Kostic <dkostic@amazon.com>